Redirect log/stderr to buffers when in console
Console use can be tough because of the logging. This change redirects
log and stderr to two respective buffers, that can be accessed via

  lastlog()

and

  lasterr()

respectively. Both methods accept a number for the length of the tail.
`lastlog` also understands `prefix=str` and `level=str` kwargs in order
to filter the output by logger-prefix and loglevel.