+ contract deployment + documentation + minor refactoring
Added code to deploy a smart contract that supports the usage of
constructor parameters. The compilation and deployment of the contract
need to be done together because of library symbols resolution and
constructor parameters encoding.

Renamed AbiContract to ContractProxy to reflect the fact that it will do
a proxy call. Simplified the implementation by factoring out the class
that represents a method.